This is the updated version of the code
