--- AzureADSample/AzureADSample/ViewModel/MainViewModel.cs
@@ -1,8 +1,0 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
-
-namespace AzureADSample.ViewModel;
-
-public partial class MainViewModel : ObservableObject
-{
-    public MainViewModel() { }
-}
